+++ CHANGELOG.md
@@ -74,7 +74,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Document type flag: `--documentType` (`-t`)
 * Unnecessary fixture playground
 
-## [0.0.2]
+## [0.0.2] - 2020-12-18
 ### Added
 * Create command
 * Play command